# platform/swaggo/swagger.yaml
basePath: /flighthours/api/v1
definitions:
...
  handlers.AircraftFamilyListResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      aircraft_families:
        items:
          $ref: '#/definitions/handlers.AircraftFamilyResponse'
        type: array
      total:
        type: integer
    type: object
  handlers.AircraftFamilyResponse:
    properties:
      family:
        type: string
      manufacturer:
        type: string
    type: object
  handlers.AircraftModelListResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      aircraft_models:
        items:
          $ref: '#/definitions/handlers.AircraftModelResponse'
        type: array
      total:
        type: integer
    type: object
  handlers.AircraftModelResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      aircraft_type_name:
        type: string
      engine_type_name:
        type: string
      id:
        type: string
      model_name:
        type: string
      status:
        type: boolean
    type: object
...
  handlers.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@champion19.com
    name: Champion19 Support
  description: API para gestión de horas de vuelo de pilotos y tripulantes
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Flighthours API
  version: "1.0"
paths:
  /aircraft-families/{family}:
    get:
      description: Returns aircraft family information (family and manufacturer) for
        a specific family (HU32)
      parameters:
      - description: Aircraft Family (e.g., A320, B737)
        in: path
        name: family
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AircraftFamilyListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get aircraft family by family name
      tags:
      - Aircraft Families
  /aircraft-models:
    get:
      description: Returns a list of all aircraft models with optional engine type
        filter
      parameters:
      - description: Filter by engine type (e.g., JET, TUR)
        in: query
        name: engine_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AircraftModelListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List all aircraft models
      tags:
      - Aircraft Models
  /aircraft-models/{id}:
    get:
      consumes:
      - application/json
      description: Returns aircraft model information by ID (accepts both UUID and
        obfuscated ID)
      parameters:
      - description: Aircraft Model ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get aircraft model by ID
      tags:
      - Aircraft Models
  /aircraft-models/{id}/activate:
    patch:
      consumes:
      - application/json
      description: Sets aircraft model status to active (HU42 - accepts obfuscated
        ID only)
      parameters:
      - description: Aircraft Model ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Activate aircraft model
      tags:
      - Aircraft Models
  /aircraft-models/{id}/deactivate:
    patch:
      consumes:
      - application/json
      description: Sets aircraft model status to inactive (HU41 - accepts obfuscated
        ID only)
      parameters:
      - description: Aircraft Model ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Deactivate aircraft model
      tags:
      - Aircraft Models
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token (format: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"